현재 오브젝트를 중심으로 하는 단순 Collider만 존재 / Collider를 Mesh로 옮기고 Bone 행렬들을 활용할 예정 - 실제로 Bone과 매칭된 Collider를 사용할 일은 별로 없을듯, 단순 Collider로 충돌체크 충분 - 다만 무기의 장착에 있어 Bone 행렬을 활용할 필요가 있다고 판단, 할줄 알아야 함

+++ Senier_Project/Senier_Project/Common/Collider.cpp
@@ -201,7 +201,7 @@ void Collider::BuildMesh(ID3D12Device* pd3dDevice, ID3D12GraphicsCommandList* pd
 
 	const UINT positionBufferByteSize = (UINT)Positions.size() * sizeof(XMFLOAT3);
 	const UINT NormalBufferByteSize = (UINT)Normal.size() * sizeof(XMFLOAT3);
-	const UINT indexBufferByteSize = (UINT)Indices.size() * sizeof(std::uint_fast32_t);
+	const UINT indexBufferByteSize = (UINT)Indices.size() * sizeof(std::uint_fast16_t);
 
 
 	CreateVertexBuffer(pd3dDevice, pd3dCommandList,
